01_PROYECTOS/01_ARNALDO_AGENCIA/backends/system-ia-agentes/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.6.0
requests>=2.31.0
python-multipart>=0.0.9
openai>=1.13.3
Pillow>=10.0.0
google-genai>=1.0.0
supabase>=2.3.0
cloudinary>=1.36.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
